feat(EG-779): enable superusers to view and edit labs
